Move core headers to their own directory
