+++ Headers/DebugServer2/Architecture/ARM/SoftwareSingleStep.h
@@ -12,7 +12,7 @@
 #define __DebugServer2_Architecture_ARM_SoftwareSingleStep_h
 
 #include "DebugServer2/Architecture/CPUState.h"
-#include "DebugServer2/BreakpointManager.h"
+#include "DebugServer2/Core/BreakpointManager.h"
 #include "DebugServer2/Target/Process.h"
 
 namespace ds2 {
