+++ Headers/DebugServer2/Core/ErrorCodes.h
