+++ Headers/DebugServer2/Types.h
@@ -12,9 +12,9 @@
 #define __DebugServer2_Types_h
 
 #include "DebugServer2/Base.h"
-#include "DebugServer2/CPUTypes.h"
 #include "DebugServer2/Constants.h"
-#include "DebugServer2/ErrorCodes.h"
+#include "DebugServer2/Core/CPUTypes.h"
+#include "DebugServer2/Core/ErrorCodes.h"
 
 #include <cstdint>
 #include <cstdlib>
