+++ Sources/Host/Linux/ARM/PTraceARM.cpp
@@ -9,7 +9,7 @@
 //
 
 #include "DebugServer2/Host/Linux/PTrace.h"
-#include "DebugServer2/HardwareBreakpointManager.h"
+#include "DebugServer2/Core/HardwareBreakpointManager.h"
 #include "DebugServer2/Host/Linux/ExtraWrappers.h"
 #include "DebugServer2/Host/Platform.h"
 #include "DebugServer2/Utils/Log.h"
